Fixed all parsers and most tests.
